refactor(StorageSize): Improve and organize code
